requirements.txt
telethon==1.36.0
python-telegram-bot==20.7
groq==0.11.0
python-dotenv==1.0.0
aiosqlite==0.20.0
apscheduler==3.10.4
pytz==2024.1
